requirements.txt
django==4.0.7
djangorestframework==3.14.0
